requirements.txt
python-telegram-bot==20.7
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
nest-asyncio>=1.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

